move objective eval and solution writing into the MPS library
